+++ CHANGELOG.md
@@ -13,6 +13,7 @@ and since Bareos version 20 this project adheres to [Semantic Versioning](https:
 - VMware Plugin: Backup and Restore NVRAM [PR #1733]
 - doc: add backtick around *.?* description  [PR #1757]
 - python-bareos: use TLS-PSK from core ssl module (available since Python >= 3.13) [PR #1764]
+- docs: improve handling of ResourceItem descriptions [PR #1766]
 
 ### Documentation
 - docs: fix Pool explanation for migration jobs [PR #1735]
@@ -448,4 +449,5 @@ and since Bareos version 20 this project adheres to [Semantic Versioning](https:
 [PR #1749]: https://github.com/bareos/bareos/pull/1749
 [PR #1757]: https://github.com/bareos/bareos/pull/1757
 [PR #1764]: https://github.com/bareos/bareos/pull/1764
+[PR #1766]: https://github.com/bareos/bareos/pull/1766
 [unreleased]: https://github.com/bareos/bareos/tree/master